Break up Data.ScopeGraph and reorg semantic-scope-graph.
