+++ semantic-python/src/Language/Python.hs
@@ -4,14 +4,14 @@ module Language.Python
 , Language.Python.Grammar.tree_sitter_python
 ) where
 
+import qualified AST.Unmarshal as TS
 import           Data.Proxy
 import qualified Language.Python.AST as Py
+import qualified Language.Python.Grammar (tree_sitter_python)
 import           Language.Python.ScopeGraph
 import qualified Language.Python.Tags as PyTags
-import           ScopeGraph.Convert
+import           Scope.Graph.Convert
 import qualified Tags.Tagging.Precise as Tags
-import qualified Language.Python.Grammar (tree_sitter_python)
-import qualified AST.Unmarshal as TS
 
 newtype Term a = Term { getTerm :: Py.Module a }
