+++ semantic-python/src/Language/Python/ScopeGraph.hs
@@ -25,6 +25,9 @@ import qualified Analysis.Name as Name
 import           AST.Element
 import           Control.Effect.Fresh
 import           Control.Effect.ScopeGraph
+import qualified Control.Effect.ScopeGraph.Properties.Declaration as Props
+import qualified Control.Effect.ScopeGraph.Properties.Function as Props
+import qualified Control.Effect.ScopeGraph.Properties.Reference as Props
 import           Control.Lens (set, (^.))
 import           Data.Foldable
 import           Data.Maybe
@@ -36,10 +39,7 @@ import           GHC.Records
 import           GHC.TypeLits
 import qualified Language.Python.AST as Py
 import           Language.Python.Patterns
-import           ScopeGraph.Convert (Result (..), complete, todo)
-import qualified ScopeGraph.Properties.Declaration as Props
-import qualified ScopeGraph.Properties.Function as Props
-import qualified ScopeGraph.Properties.Reference as Props
+import           Scope.Graph.Convert (Result (..), complete, todo)
 import           Source.Loc
 import           Source.Span (span_)
