+++ semantic-scope-graph/src/Control/Carrier/Sketch/ScopeGraph.hs
@@ -26,6 +26,7 @@ import           Control.Carrier.Fresh.Strict
 import           Control.Carrier.Reader
 import           Control.Carrier.State.Strict
 import           Control.Effect.ScopeGraph (ScopeGraphEff (..))
+import qualified Control.Effect.ScopeGraph.Properties.Declaration as Props
 import           Control.Monad.IO.Class
 import           Data.Bifunctor
 import qualified Data.List.NonEmpty as NonEmpty
@@ -33,7 +34,6 @@ import           Data.Module
 import           Data.ScopeGraph (ScopeGraph)
 import qualified Data.ScopeGraph as ScopeGraph
 import           Data.Semilattice.Lower
-import qualified ScopeGraph.Properties.Declaration as Props
 import           Source.Span
 import qualified System.Path as Path
