+++ semantic-scope-graph/src/Control/Effect/ScopeGraph/Properties/Declaration.hs
@@ -5,7 +5,7 @@
 -- | The 'Declaration' record type is used by the 'Control.Effect.Sketch' module to keep
 -- track of the parameters that need to be passed when establishing a new declaration.
 -- That is to say, it is a record type primarily used for its selector names.
-module ScopeGraph.Properties.Declaration
+module Control.Effect.ScopeGraph.Properties.Declaration
   ( Declaration (..)
   ) where
