+++ semantic-scope-graph/src/Control/Effect/ScopeGraph/Properties/Function.hs
@@ -5,7 +5,7 @@
 -- | The 'Function' record type is used by the 'Control.Effect.Sketch' module to keep
 -- track of the parameters that need to be passed when establishing a new declaration.
 -- That is to say, it is a record type primarily used for its selector names.
-module ScopeGraph.Properties.Function
+module Control.Effect.ScopeGraph.Properties.Function
   ( Function (..)
   ) where
